Expanded genutils.sequences into full module (with prior functionality moved into seqops.py submodule), updated dependencies
